legacy_python/requirements.txt
google-generativeai
google-cloud-firestore
google-cloud-storage
google-cloud-secret-manager
fastapi
uvicorn
jinja2
python-multipart
google-analytics-data
requests
python-dotenv
flask
gunicorn
functions-framework
google-cloud-scheduler
supabase
google-cloud-aiplatform
